Refactoring task system
Switch dispatcher tasks to asio io_service and scheduler tasks to asio deadline_timers.
It makes code more simple and helps to avoid unnecessary thread-context switches.

+++ src/decay.h
@@ -31,7 +31,7 @@ class Decay
 	private:
 		void checkDecay();
 
-		uint32_t eventId {0};
+		uint64_t eventId {0};
 		std::map<int64_t, std::vector<Item*>> decayMap;
 };
